+++ e2e-test/cypress/support/commands.js
@@ -212,6 +212,27 @@ Cypress.Commands.add("updateProjectPermissions", (projectId, intent, identity) =
     .then(body => Promise.resolve(body.data));
 });
 
+Cypress.Commands.add("updateSubprojectPermissions", (projectId, subprojectId, intent, identity) => {
+  cy.request({
+    url: `${baseUrl}/api/subproject.intent.grantPermission`,
+    method: "POST",
+    headers: {
+      Authorization: `Bearer ${token}`
+    },
+    body: {
+      apiVersion: "1.0",
+      data: {
+        projectId: projectId,
+        subprojectId: subprojectId,
+        identity: identity,
+        intent: intent
+      }
+    }
+  })
+    .its("body")
+    .then(body => Promise.resolve(body.data));
+});
+
 Cypress.Commands.add("grantUserPermissions", (userId, intent, identity) => {
   cy.request({
     url: `${baseUrl}/api/user.intent.grantPermission`,